feat(hero): add dropdown
